requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
Werkzeug==2.3.7
Jinja2==3.1.2
python-dateutil==2.8.2
requests==2.31.0
Pillow==11.0.0
gunicorn==21.2.0
python-dotenv==1.0.0

# Testing Framework
pytest>=8.0
pytest-cov>=5.0
pytest-mock>=3.10
pytest-xdist>=3.0

# Coverage
coverage[toml]>=7.0

# Caching
Flask-Caching==2.0.2

# Security
bcrypt==4.0.1